visitorの修正
